refactor: プロフィール取得をサービス層に統一
