+++ app/routes/admin/layout.tsx
@@ -5,30 +5,26 @@ import { Outlet, redirect } from "react-router";
 
 import { AdminLayout } from "~/components/layouts/adminLayout/adminLayout";
 import { SidebarSignOutButton } from "~/components/ui/sidebar";
-import { createServerSupabaseClient } from "~/services/supabase/client.server";
-import { getProfileByUserId } from "~/services/supabase/profiles";
+import { getProfileByUserId } from "~/services/profiles/get.server";
 import { hasModeratorPermission } from "~/utils/permissions";
 
 import type { Route } from "./+types/layout";
 
 export const loader = async (args: Route.LoaderArgs) => {
-  const supabase = createServerSupabaseClient(args);
   const auth = await getAuth(args);
 
   if (!auth.isAuthenticated) {
     throw redirect("/");
   }
 
-  const profileResponse = auth.userId
-    ? await getProfileByUserId(supabase, auth.userId)
-    : null;
+  const profileResponse = await getProfileByUserId(args, auth.userId);
 
-  if (!profileResponse?.data) {
+  if (!profileResponse.success) {
     throw redirect("/");
   }
 
   const userProfile = profileResponse.data;
-  const permissionLevel = userProfile.roles?.permission_level;
+  const permissionLevel = userProfile.roles.permission_level;
 
   if (!hasModeratorPermission(permissionLevel)) {
     throw redirect("/unauthorized");
